requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.7
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
requests==2.31.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0